+++ docs/release_notes.md
@@ -13,7 +13,7 @@
 
 ### ✨ New Functionality
 
--
+- [Orchestration] Added support for [transforming a JSON output into an entity](https://sap.github.io/ai-sdk/docs/java/orchestration/chat-completion#json_schema)
 
 ### 📈 Improvements